[refactor] #14 - 컴포넌트별 분리
